+++ src/main/java/com/axibase/tsd/driver/jdbc/DriverConstants.java
@@ -10,7 +10,7 @@ public final class DriverConstants {
 	public static final String DATABASE_PRODUCT_NAME = "Axibase";
 	public static final String DATABASE_PRODUCT_VERSION = "Axibase Time Series Database";
 	public static final String JDBC_DRIVER_NAME = "ATSD JDBC driver";
-	public static final String JDBC_DRIVER_VERSION_DEFAULT = "1.4.5";
+	public static final String JDBC_DRIVER_VERSION_DEFAULT = "1.4.6";
 	public static final int    DRIVER_VERSION_MAJOR_DEFAULT = 1;
 	public static final int    DRIVER_VERSION_MINOR_DEFAULT = 3;
 
@@ -67,6 +67,8 @@ public final class DriverConstants {
 	public static final int API_FIND_METRICS_STATISTICS_RENAME_REV = 19437;
 	public static final int API_METRIC_STATISTICS_RENAME_REV = 20295;
 	public static final int META_COLUMNS_REVISION = 20356;
+	public static final int META_COLUMNS_SHORT_REVISION = 20377;
+	public static final int META_COLUMNS_CREATION_TIME_REVISION = 20367;
 	public static final String REVISION_LINE = "Revision";
 
 	public static final String CONN_KEEP_ALIVE = "Keep-Alive";