+++ src/main/java/com/axibase/tsd/driver/jdbc/enums/DefaultTable.java
@@ -3,19 +3,30 @@ package com.axibase.tsd.driver.jdbc.enums;
 import com.axibase.tsd.driver.jdbc.intf.MetadataColumnDefinition;
 
 public enum DefaultTable {
-    ATSD_SERIES("SELECT metric, entity, tags.collector, " +
-            "tags.host, datetime, time, value FROM atsd_series WHERE metric = 'gc_time_percent' " +
-            "AND entity = 'atsd' AND datetime >= now - 5*MINUTE ORDER BY datetime DESC LIMIT 10"),
-    ATSD_METRIC(prepareRemark("atsd_metric", MetricColumn.values(), "metric = 'gc_time_percent'")),
-    ATSD_ENTITY(prepareRemark("atsd_entity", EntityColumn.values(), "entity = 'atsd'"));
+    ATSD_SERIES {
+        @Override
+        public String getRemark(boolean shortNames) {
+            return "SELECT metric, entity, tags.collector, " +
+                    "tags.host, datetime, time, value FROM atsd_series WHERE metric = 'gc_time_percent' " +
+                    "AND entity = 'atsd' AND datetime >= now - 5*MINUTE ORDER BY datetime DESC LIMIT 10";
+        }
+    },
+    ATSD_METRIC {
+        @Override
+        public String getRemark(boolean shortNames) {
+            return prepareRemark("atsd_metric", MetricColumn.values(), "metric = 'gc_time_percent'", shortNames);
+        }
+    },
+    ATSD_ENTITY {
+        @Override
+        public String getRemark(boolean shortNames) {
+            return prepareRemark("atsd_entity", EntityColumn.values(), "entity = 'atsd'", shortNames);
+        }
+    };
 
     public final String tableName = name().toLowerCase();
 
-    public final String remark;
-
-    DefaultTable(String remark) {
-        this.remark = remark;
-    }
+    public abstract String getRemark(boolean shortNames);
 
     public static boolean isDefaultTable(String table) {
         return ATSD_SERIES.tableName.equalsIgnoreCase(table)
@@ -23,10 +34,11 @@ public enum DefaultTable {
                 || ATSD_METRIC.tableName.equalsIgnoreCase(table);
     }
 
-    private static String prepareRemark(String tableName, MetadataColumnDefinition[] columns, String where) {
+    private static String prepareRemark(String tableName, MetadataColumnDefinition[] columns, String where, boolean shortNames) {
         final StringBuilder buffer = new StringBuilder("SELECT ");
         for (MetadataColumnDefinition column : columns) {
-            buffer.append(column.getColumnNamePrefix()).append(", ");
+            String prefix = shortNames ? column.getShortColumnNamePrefix() : column.getColumnNamePrefix();
+            buffer.append(prefix).append(", ");
         }
         buffer.setLength(buffer.length() - 2);
         return buffer.append(" FROM ").append(tableName)