+++ src/main/java/com/axibase/tsd/driver/jdbc/enums/EntityColumn.java
@@ -1,15 +1,15 @@
 package com.axibase.tsd.driver.jdbc.enums;
 
-import static com.axibase.tsd.driver.jdbc.enums.AtsdType.BOOLEAN_DATA_TYPE;
 import com.axibase.tsd.driver.jdbc.intf.MetadataColumnDefinition;
 import com.axibase.tsd.driver.jdbc.util.AtsdColumn;
 import lombok.Getter;
+import org.apache.commons.lang3.ArrayUtils;
 
-import static com.axibase.tsd.driver.jdbc.enums.AtsdType.STRING_DATA_TYPE;
+import static com.axibase.tsd.driver.jdbc.enums.AtsdType.*;
 
 @Getter
 public enum EntityColumn implements MetadataColumnDefinition {
-
+    CREATION_TIME(AtsdColumn.ENTITY_CREATION_TIME, BIGINT_DATA_TYPE),
     ENABLED(AtsdColumn.ENTITY_ENABLED, BOOLEAN_DATA_TYPE),
     GROUPS(AtsdColumn.ENTITY_GROUPS, STRING_DATA_TYPE),
     INTERPOLATE(AtsdColumn.ENTITY_INTERPOLATE, STRING_DATA_TYPE),
@@ -27,6 +27,8 @@ public enum EntityColumn implements MetadataColumnDefinition {
         this.type = type;
     }
 
+    private static final int PREFIX_LENGTH = "entity.".length();
+
     @Override
     public String getNullableAsString() {
         return NULLABLE_AS_STRING[nullable];
@@ -35,6 +37,16 @@ public enum EntityColumn implements MetadataColumnDefinition {
     @Override
     public AtsdType getType(AtsdType metricType) {
         return type;
+    }
+
+    @Override
+    public String getShortColumnNamePrefix() {
+        return columnNamePrefix.substring(PREFIX_LENGTH);
+    }
+
+    public static EntityColumn[] values(boolean withCreationTime) {
+        final EntityColumn[] values = EntityColumn.values();
+        return withCreationTime ? values : ArrayUtils.removeElements(values, CREATION_TIME);
     }
 
 }