+++ src/main/java/com/axibase/tsd/driver/jdbc/ext/AtsdMeta.java
@@ -178,7 +178,7 @@ public class AtsdMeta extends MetaImpl {
 			}
 			return result;
 		} catch (SQLDataException | SQLFeatureNotSupportedException e) {
-			log.error("[execute] error", e.getMessage());
+			log.error("[execute] error {}", e.getMessage());
 			throw e;
 		} catch (final RuntimeException e) {
 			log.error("[execute] error", e);
@@ -298,7 +298,7 @@ public class AtsdMeta extends MetaImpl {
 	@SneakyThrows({SQLDataException.class, SQLFeatureNotSupportedException.class})
 	public ExecuteResult prepareAndExecute(StatementHandle statementHandle, String query, long maxRowCount,
 										   int maxRowsInFrame, PrepareCallback callback) {
-		final long limit = maxRowCount < 0 ? 0 : maxRowCount;
+		final long limit = Math.max(0, maxRowCount);
 		log.trace("[prepareAndExecute] handle: {} maxRowCount: {} query: {}", statementHandle, limit, query);
 		try {
 			final AvaticaStatement statement = (AvaticaStatement) callback.getMonitor();
@@ -325,7 +325,7 @@ public class AtsdMeta extends MetaImpl {
 			callback.execute();
 			return result;
 		} catch (SQLDataException | SQLFeatureNotSupportedException e) {
-			log.error("[prepareAndExecute] error", e.getMessage());
+			log.error("[prepareAndExecute] error {}", e.getMessage());
 			throw e;
 		} catch (final AtsdRuntimeException e) {
 			log.error("[prepareAndExecute] error", e);
@@ -361,7 +361,7 @@ public class AtsdMeta extends MetaImpl {
 			}
 			return new ExecuteBatchResult(updateCounts);
 		} catch (SQLDataException | SQLFeatureNotSupportedException e) {
-			log.error("[prepareAndExecuteBatch] error", e.getMessage());
+			log.error("[prepareAndExecuteBatch] error {}", e.getMessage());
 			throw e;
 		} catch (final RuntimeException e) {
             log.error("[prepareAndExecuteBatch] error", e);
@@ -394,7 +394,7 @@ public class AtsdMeta extends MetaImpl {
 			long[] updateCounts = updateCount == 0 ? new long[parameterValueBatch.size()] : converter.getCommandCounts();
 			return new ExecuteBatchResult(updateCounts);
 		} catch (SQLDataException | SQLFeatureNotSupportedException e) {
-			log.error("[executeBatch] error", e.getMessage());
+			log.error("[executeBatch] error {}", e.getMessage());
 			throw e;
 		} catch (final RuntimeException e) {
 			log.error("[executeBatch] error", e);
@@ -555,7 +555,7 @@ public class AtsdMeta extends MetaImpl {
 				table, "TABLE", remark);
 	}
 
-	private String generateTableRemark(String table) {
+	private String generateTableRemark(String table, boolean addCreationTime) {
 		StringBuilder buffer = new StringBuilder("SELECT");
         for (DefaultColumn column : DefaultColumn.values()) {
             if (column.ordinal() != 0) {
@@ -564,26 +564,40 @@ public class AtsdMeta extends MetaImpl {
             buffer.append(' ').append(column.getColumnNamePrefix());
         }
 		if (atsdConnectionInfo.metaColumns()) {
-		    appendMetaColumns(buffer, EntityColumn.values());
-		    appendMetaColumns(buffer, MetricColumn.values());
+		    appendMetaColumns(buffer, EntityColumn.values(addCreationTime));
+		    appendMetaColumns(buffer, MetricColumn.values(addCreationTime));
 		}
 		return buffer.append(" FROM \"").append(table).append("\" LIMIT 1").toString();
 	}
 
+	/**
+	 * Append metric and entity metadata columns to series requests. This method must not be used while processing meta tables.
+	 * @param buffer StringBuilder to append to
+	 * @param values columns to append
+	 */
 	private static void appendMetaColumns(StringBuilder buffer, MetadataColumnDefinition[] values) {
         for (MetadataColumnDefinition column : values) {
             buffer.append(", ").append(column.getColumnNamePrefix());
         }
     }
 
+    private boolean useShortColumnNamesInMetaTables() {
+		return getAtsdRevision() >= META_COLUMNS_SHORT_REVISION;
+	}
+
+	private boolean shouldAddCreationTimeMetaColumns() {
+		return getAtsdRevision() >= META_COLUMNS_CREATION_TIME_REVISION;
+	}
+
 	private List<Object> receiveTables(AtsdConnectionInfo connectionInfo, String pattern) {
 		final List<Object> metricList = new ArrayList<>();
 		final List<String> metricMasks = connectionInfo.tables();
+		final boolean metadataColumnsShort = useShortColumnNamesInMetaTables();
 		for (DefaultTable defaultTable : DefaultTable.values()) {
 			for (String metricMask : metricMasks) {
 				if (WildcardsUtil.wildcardMatch(defaultTable.tableName, metricMask)
 						&& WildcardsUtil.wildcardMatch(defaultTable.tableName, pattern)) {
-					metricList.add(generateMetaTable(defaultTable.tableName, defaultTable.remark));
+					metricList.add(generateMetaTable(defaultTable.tableName, defaultTable.getRemark(metadataColumnsShort)));
 				}
 			}
 		}
@@ -597,8 +611,9 @@ public class AtsdMeta extends MetaImpl {
             }
         }
 
+        final boolean addCreationTime = shouldAddCreationTimeMetaColumns();
         for (String metricName : metricNamesToTypes.keySet()) {
-            metricList.add(generateMetaTable(metricName, generateTableRemark(metricName)));
+            metricList.add(generateMetaTable(metricName, generateTableRemark(metricName, addCreationTime)));
         }
 
 		return metricList;
@@ -753,7 +768,8 @@ public class AtsdMeta extends MetaImpl {
         final List<String> metricMasks = atsdConnectionInfo.tables();
 		if (!metricMasks.isEmpty()) {
 			final String colNamePattern = columnNamePattern.s;
-			final List<MetadataColumnDefinition> columns = filterColumns(colNamePattern, atsdConnectionInfo.metaColumns());
+			final boolean addCreationTime = shouldAddCreationTimeMetaColumns();
+			final List<MetadataColumnDefinition> columns = filterColumns(colNamePattern, atsdConnectionInfo.metaColumns(), addCreationTime);
 			final List<Object> columnData = new ArrayList<>();
 			final boolean underscoreAsLiteral = atsdConnectionInfo.disableUnderscoreInColumns();
 			final String pattern = StringUtils.isBlank(schemaPattern.s) ? tableNamePattern.s : schemaPattern.s + '.' + tableNamePattern.s;
@@ -770,20 +786,22 @@ public class AtsdMeta extends MetaImpl {
 				}
 			}
 			final boolean odbcCompatible = atsdConnectionInfo.odbc2Compatibility();
+			final boolean useShortNames = useShortColumnNamesInMetaTables();
 			for (Map.Entry<String, AtsdType> entry : tableNamesAndValueTypes.entrySet()) {
 				final String tableName = entry.getKey();
 				final AtsdType metricValueType = entry.getValue();
 				final List<MetadataColumnDefinition> columnDefinitions;
 				if (DefaultTable.ATSD_METRIC.tableName.equals(tableName)) {
-					columnDefinitions = filterMetaTablesColumns(colNamePattern, MetricColumn.values());
+					columnDefinitions = filterMetaTablesColumns(colNamePattern, MetricColumn.values(addCreationTime), useShortNames);
 				} else if (DefaultTable.ATSD_ENTITY.tableName.equals(tableName)) {
-					columnDefinitions = filterMetaTablesColumns(colNamePattern, EntityColumn.values());
+					columnDefinitions = filterMetaTablesColumns(colNamePattern, EntityColumn.values(addCreationTime), useShortNames);
 				} else {
 					columnDefinitions = columns;
 				}
 				int position = 1;
+				final boolean cutMetaPrefix = useShortNames && columnDefinitions != columns;
 				for (MetadataColumnDefinition column : columnDefinitions) {
-					columnData.add(createColumnMetaData(column, tableName, metricValueType, position, odbcCompatible));
+					columnData.add(createColumnMetaData(column, tableName, metricValueType, position, odbcCompatible, cutMetaPrefix));
 					++position;
 				}
 				appendFilteredTagsColumns(tableName, colNamePattern, columnData, odbcCompatible, metricValueType);
@@ -798,20 +816,22 @@ public class AtsdMeta extends MetaImpl {
 		return createEmptyResultSet(AtsdMetaResultSets.AtsdMetaColumn.class);
 	}
 
-	private void appendFilteredTagsColumns(String tableName, String colNamePattern, List<Object> columnData, boolean odbcCompatible, AtsdType metricValueType) {
+	private void appendFilteredTagsColumns(String tableName, String colNamePattern, List<Object> columnData,
+										   boolean odbcCompatible, AtsdType metricValueType) {
         if (DefaultTable.isDefaultTable(tableName) || !maybeTagColumnPattern(colNamePattern)) {
             return;
         }
+        final boolean useShortNames = false; // don't use in non-default tables
         final Set<String> tags = getTags(tableName);
         int position = columnData.size() + 1;
         if (tags.isEmpty() && StringUtils.startsWith(colNamePattern, TagColumn.PREFIX) && !WildcardsUtil.hasWildcards(colNamePattern)) {
             final TagColumn column = new TagColumn(StringUtils.substringAfter(colNamePattern, TagColumn.PREFIX));
-            columnData.add(createColumnMetaData(column, tableName, metricValueType, position, odbcCompatible));
+            columnData.add(createColumnMetaData(column, tableName, metricValueType, position, odbcCompatible, useShortNames));
         } else {
             for (String tag : tags) {
                 final TagColumn column = new TagColumn(tag);
                 if (WildcardsUtil.wildcardMatch(column.getColumnNamePrefix(), colNamePattern)) {
-                    columnData.add(createColumnMetaData(column, tableName, metricValueType, position, odbcCompatible));
+                    columnData.add(createColumnMetaData(column, tableName, metricValueType, position, odbcCompatible, useShortNames));
                     ++position;
                 }
             }
@@ -839,32 +859,34 @@ public class AtsdMeta extends MetaImpl {
                 || pattern.startsWith(MetricColumn.TAGS.getColumnNamePrefix());
 	}
 
-	private static List<MetadataColumnDefinition> filterColumns(String columnPattern, boolean showMetaColumns) {
+	private static List<MetadataColumnDefinition> filterColumns(String columnPattern, boolean showMetaColumns, boolean addCreationTime) {
 		List<MetadataColumnDefinition> result = new ArrayList<>();
 		for (DefaultColumn column : DefaultColumn.values()) {
-			filterColumn(columnPattern, column, result);
+			filterColumn(columnPattern, column.getColumnNamePrefix(), column, result);
 		}
 		if (showMetaColumns) {
-			for (EntityColumn column : EntityColumn.values()) {
-				filterColumn(columnPattern, column, result);
+			for (EntityColumn column : EntityColumn.values(addCreationTime)) {
+				filterColumn(columnPattern, column.getColumnNamePrefix(), column, result);
 			}
-			for (MetricColumn column : MetricColumn.values()) {
-				filterColumn(columnPattern, column, result);
+			for (MetricColumn column : MetricColumn.values(addCreationTime)) {
+				filterColumn(columnPattern, column.getColumnNamePrefix(), column, result);
 			}
 		}
 		return result;
 	}
 
-	private static List<MetadataColumnDefinition> filterMetaTablesColumns(String columnPattern, MetadataColumnDefinition[] unfiltered) {
+	private static List<MetadataColumnDefinition> filterMetaTablesColumns(String columnPattern, MetadataColumnDefinition[] unfiltered, boolean useShortNames) {
 		List<MetadataColumnDefinition> result = new ArrayList<>();
 		for (MetadataColumnDefinition columnDefinition : unfiltered) {
-			filterColumn(columnPattern, columnDefinition, result);
+			final String columnNamePrefix = useShortNames ? columnDefinition.getShortColumnNamePrefix() : columnDefinition.getColumnNamePrefix();
+			filterColumn(columnPattern, columnNamePrefix, columnDefinition, result);
 		}
 		return result;
 	}
 
-	private static void filterColumn(String columnPattern, MetadataColumnDefinition column, final List<MetadataColumnDefinition> columns) {
-		if (WildcardsUtil.wildcardMatch(column.getColumnNamePrefix(), columnPattern)) {
+	private static void filterColumn(String columnPattern, String columnNamePrefix, MetadataColumnDefinition column,
+									 final List<MetadataColumnDefinition> columns) {
+		if (WildcardsUtil.wildcardMatch(columnNamePrefix, columnPattern)) {
 			columns.add(column);
 		}
 	}
@@ -896,14 +918,15 @@ public class AtsdMeta extends MetaImpl {
 		return baseMetricUrl + "/" + encodedMetric + "/series";
 	}
 
-	private AtsdMetaResultSets.AtsdMetaColumn createColumnMetaData(MetadataColumnDefinition column, String table, AtsdType valueType, int ordinal, boolean odbcCompatible) {
+	private AtsdMetaResultSets.AtsdMetaColumn createColumnMetaData(MetadataColumnDefinition column, String table, AtsdType valueType,
+																   int ordinal, boolean odbcCompatible, boolean useShortNames) {
 		final AtsdType columnType = column.getType(valueType).getCompatibleType(odbcCompatible);
 		return new AtsdMetaResultSets.AtsdMetaColumn(
 				odbcCompatible,
 				atsdConnectionInfo.catalog(),
 				atsdConnectionInfo.schema(),
 				table,
-				column.getColumnNamePrefix(),
+				useShortNames ? column.getShortColumnNamePrefix() : column.getColumnNamePrefix(),
 				columnType,
 				column.getNullable(),
 				ordinal,